Use the internal download() function rather than curl
